feat(minesweeper): add
This was a fun project to work on for me. It was nostalgic for me since minesweeper is one of the only game I've played out of MS classic games.

This commit contains basic game play features as well as couple additional features for classic minesweeper.

This was built on `Angular 7` bootstrapped by `angular-cli`. Other libraries that `angular-cli` brings in by default mainly for corporate-level applications were left as-is. `rxjs` was used to mimic store design for managing the board (ref. `board.service.ts`). Basic styling, including icons, comes from `angular-material`. Although, I was trying to stay with JS implementations, `lodash` was pulled in for proper cloning of each cell (square) class instances as well as to minimize mutations as much as possible for the given timeline.

In terms of styling, I focused mostly on the board and squares contained rather than overall application.

Here are couple additional features and assumptions I've made while implmenting:

- Count of mines or available flags added to board toolbar
- Face button to restart the game added to board toolbar
- Timer added to board toolbar
	- Max time is set to 59 minutes and 59 seconds
	- Game ends when time runs out
- Multiple levels
	- Referenced classic minesweeper game configs (10, 40, 99 mines)
- When opening all squares around the zero square, flagged squares does not open automatically
- User also wins if closed squares left on the board are **only** mines underneath

Also, since I've played minesweeper quite intensly back in college, there are couple features I would've like to implement in this initial commit that I wasn't able to get to due to time constraint:

- Populate the grid upon first click of a square
	- This ensures user will not click a mine on their first choice
- Double click on a square checks srounding squares and besides flagged or already opened squares does not contain mines, game automatically open those